week5-spatial-data-integration/requirements.txt
# Week 5 Spatial Data Integration Platform Requirements

# Core Dependencies
requests>=2.31.0
psycopg2-binary>=2.9.7
pandas>=2.0.3
numpy>=1.24.3
geopandas>=0.13.2

# Database & Spatial
SQLAlchemy>=2.0.19
PostGIS>=3.0.0  # PostgreSQL extension
Shapely>=2.0.1
Fiona>=1.9.4

# Web Scraping & APIs
beautifulsoup4>=4.12.2
feedparser>=6.0.10
urllib3>=2.0.4
html5lib>=1.1

# Data Processing & Analysis
scipy>=1.11.1
scikit-learn>=1.3.0
geopy>=2.3.0

# Interactive Dashboard
Flask>=2.3.2
Jinja2>=3.1.2

# Environmental Data Processing
rasterio>=1.3.8
xarray>=2023.7.0

# Economic Data APIs (Government)
census>=0.8.19
fredapi>=0.5.0

# Monitoring & Logging
logging-tree>=1.9

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Government Data Sources Integration
# FEMA - National Flood Hazard Layer
# CAL FIRE - Fire Hazard Severity Zones
# USGS - Seismic Hazard Data
# Census Bureau - American Community Survey
# Bureau of Labor Statistics - Employment Data
# HUD - Fair Market Rent Data

# Note: All government data sources are free/public APIs
# No paid API keys required for core functionality